Implementation of AuthenticationGraph, signinscreen
